requirements.txt
boto3==1.18.39
botocore==1.21.39
bottle==0.12.19
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
jmespath==0.10.0
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
torchvision==0.9.0
torch==1.8.0
numpy==1.21.2
Pillow==8.3.2
gunicorn==20.1.0
